requirements.txt
# ============================================================================
# PRODUCTION, DEVELOPMENT & ENHANCED DEPENDENCIES (Unified)
# ============================================================================

# Telegram
aiogram==3.8.0           # Telegram Bot API
telethon==1.36.0         # Telegram client для мониторинга каналов

# HTTP & Web
aiohttp==3.9.5           # HTTP-client
tavily-python==0.3.3     # Web search API

# LLM & AI
langchain==0.3.26        # LangChain core
langchain-community==0.3.27  # LangChain community integrations
ollama==0.3.3            # Python-клиент Ollama
tiktoken==0.9.0           # Tokenizer
transformers==4.53.1     # HuggingFace Transformers
sentence-transformers==3.0.1  # Для семантических embeddings
sentencepiece==0.2.0     # Для токенизации
accelerate==1.8.1        # Для ускорения обучения
nltk==3.8.1              # Для NLP
torch==2.5.0             # PyTorch
rouge-score==0.1.2       # ROUGE метрики
bert-score==0.3.13       # Для BERTScore метрик
scikit-learn==1.3.0      # Для cosine similarity
numpy==1.26.4            # Numpy

# Database
psycopg[binary]==3.2.9   # PostgreSQL driver
psycopg-pool==3.2.6      # Connection pooling
asyncpg==0.30.0          # Async PostgreSQL driver
psycopg2-binary==2.9.10  # Sync PostgreSQL driver (обновлено)

# Configuration & Validation
pydantic==2.7.4          # Data validation (совместимо с aiogram)
pydantic-settings==2.10.1 # Settings management
python-dotenv==1.0.1     # Environment variables

# Templates & Utils
jinja2==3.1.3            # Template engine для промптов

# System monitoring
psutil==5.9.8

# Data analysis & visualization
pandas==2.0.3            # Для анализа данных
matplotlib==3.7.1         # Для визуализации
seaborn==0.12.2           # Для графиков
plotly==5.15.0            # Для интерактивных графиков

# Development tools
ruff==0.4.4              # Linter & formatter
mypy==1.10.0             # Type checker
pytest==8.2.0           # Testing framework
pytest-asyncio==0.23.6  # Async testing
coverage==7.5.1          # Code coverage
